Convert JsCore to the new fixpoint scheme.
`JsCore` is now the recursive type; `JsCoreF` is the Functor.

All names are now represented in a new wrapper type, `Name`. This avoids a lot of Stringly-ness, but also came with some pain. In paticular, it's not clear that it's worth changing the type of the keys in `Obj` to `Name`. For the time being, `Name` does not constrain the values it contains to actual legal JavaScript names, but maybe it should, in which case we would probably need to revisit that use. Most other uses are very clearly intended to be proper JS identifiers (method, function, and variable names).

Several convenience constructors are provided, possibly for the short-term only: `ident`, `binop`, `Select`. They mirror pre-exisiting functions so less wholesale changes were required.

The strangely named JsCore -> Bson extractor was pulled out and renamed to HasLiteral at the point of use.