General improvements to Free in core
This commits brings over the general purpose improvements to core that were done in the web branch (which is currently blocked on more work that needs to be done on Free in core) so that they can be available as soon as possible for projects that depend on Quasar as a library. Also to avoid conflicts with current work on Free in master.

This commit aggregation is co-authored with @wemrysi, specifically the part on the addition of the `Explain` term to the `QueryFile` algebra.