Use typelevel compiler.
The immediate consequence is the elimination of si2712fix-plugin.
A subsequent commit will exploit type literals and singleton inference
to eliminate the currently extensive reliance on shapeless's Nat.

The apparently unrelated change to move the convertError object
outside of the Errors object is actually necessary for things to
work with the typelevel options. This is life with the scala
compiler: any change at all is likely to alter the profile of
when types get forced, and you never know how it will materialize.
In fact it's very encouraging that only the one change was needed.